requirements/test.txt
-r prod.txt
tox
coverage
pytest
pytest-cov
factory_boy
Faker
